setup: file rename
